requirements.txt
# Core
streamlit>=1.32.0

# Image processing
Pillow>=10.0.0
opencv-python-headless>=4.9.0
numpy>=1.24.0

# Face detection / sticker overlay
mediapipe>=0.10.0

# PDF export
reportlab>=4.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0